+++ Changelog.md
@@ -3,7 +3,9 @@
 Language Features:
 
 Compiler Features:
+* Custom Storage Layout: the base slot expression can also be specified by constant variables.
 * ethdebug: Experimental support for instructions and source locations under EOF.
+
 
 Bugfixes: